Health and Economic Data from a Microsimulation Modeling Study of Revised Colorectal Cancer Screening Guidelines in Canada

This dataset is borrowed from the author's personal research project in a modified format. The data was generated from a microsimulation model comparing various colorectal cancer screening strategies to assess the health benefits and economic burden of improving screening outomes.

Usage

data("easyicer_data")

Format

A data frame with 828 observations on the following 8 variables.

cohort

a numeric vector

scr_age

a numeric vector

ppt_rate

a numeric vector

scr_mod

a character vector

year

a numeric vector

Cost_of_screen_x_reen_positives

a numeric vector

Cost_of_clinic_x_ow_up_protocol

a numeric vector

Health_adjusted_person_years

a numeric vector

Source

This dataset is not publicly available. The modified dataset is contained in a .csv in the GitHub repository.

Examples

data(easyicer_data)
## e.g., str(easyicer_data) ; colnames(easyicer_data) ...